spatial processing steps (not working script)
This is not a working script. Instead, this file provides a rough record of all spatial processing steps taken to compile remotely sensed predictors and extract values for those predictors to nest and random points. Python snippets for most (but not all) steps were copied and pasted into this file from the Geoprocessing Results window in ArcCatalog as completed.